feat(rar): add support for multi-volume rar archives
The RAR rchive format supports multi-volume archives. These are archives
that are split across multiple files and extracting files from them
requires reading through each of the 'volumes' until all the data is
retrieved.

For streaming, this doesn't make sense (as you could just stream all the
data in a single stream - no reason to split it) but for reading from
disk, the way this is now handled is by setting an additional field on
the Rar struct.

The Name field specifies the filename of the _first_ volume in the
multi-volume archive. This is important as this name is used to
derive the names of subsequent volumes in the archive.

Additionally, you can also set an FS field on the Rar struct. If set,
this forces the extractor to load volumes from the provided fs.FS rather
than directly via OS.