Update dependencies to avoid CVE-2025-58058 (#48)
xz is a pure golang package for reading and writing xz-compressed files.
Prior to version 0.5.14, it is possible to put data in front of an
LZMA-encoded byte stream without detecting the situation while reading
the header.

there are also others in there that had patch level updates, update
them all