refactor: make extension detection more robust
Current extension detection relies on string search
which can lead to situation where a normal file is
seen as an archive such as with the brotli extension
which is br and the Bru markup language which uses
bru.

This refactor implement the detection based on
splitting the file name on dots and then check
the artifacts base on that. While not bullet
proof it should keep the current behaviour while
avoiding sitution where close extension gets mixed.