Make extension detection more robust (#52)
* refactor: make extension detection more robust

Current extension detection relies on string search
which can lead to situation where a normal file is
seen as an archive such as with the brotli extension
which is br and the Bru markup language which uses
bru.

This refactor implement the detection based on
splitting the file name on dots and then check
the artifacts base on that. While not bullet
proof it should keep the current behaviour while
avoiding sitution where close extension gets mixed.

* refactor: remove internal utils and use filepath

All file types are now tested using equality with
their respective extension except for .tar (which
matches the original implementation).

* refactor: check tar for the exact extension not just contains

* chore: remove debug print

Co-authored-by: Matt Holt <mholt@users.noreply.github.com>

* chore: add more tests

* fix(tar): remove first element of slice as it should be the file name

---------

Co-authored-by: Matt Holt <mholt@users.noreply.github.com>